no need to recreate cert if existed

+++ gen-root-ca.sh
@@ -1,3 +1,9 @@
+# If etcd-root-ca.pem and etcd-root-ca-key.pem already exist, then skip this step
+if [ -f certs/etcd-root-ca.pem ] && [ -f certs/etcd-root-ca-key.pem ]; then
+  echo "Root CA already exists. Skipping root CA generation."
+  exit 0
+fi
+
 # Make sure bin and certs directories exist and clean
 mkdir -p bin && rm -f bin/cfssl* && rm -rf certs && mkdir -p certs
